packages/curate-common/src/curate_common/models/base.py
"""Base model with shared fields for all Cosmos DB documents."""

from __future__ import annotations

from datetime import UTC, datetime
from uuid import uuid4

from pydantic import BaseModel, Field


def _utcnow() -> datetime:
    return datetime.now(UTC)


def _new_id() -> str:
    return str(uuid4())


class DocumentBase(BaseModel):
    """Base for all Cosmos DB document types."""

    id: str = Field(default_factory=_new_id)
    created_at: datetime = Field(default_factory=_utcnow)
    updated_at: datetime = Field(default_factory=_utcnow)
    deleted_at: datetime | None = None

    model_config = {"populate_by_name": True}
