packages/curate-common/pyproject.toml
[project]
name = "curate-common"
version = "0.1.0"
description = "Shared library for Curate — config, models, database, and storage."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-cosmos>=4.14.6",
    "azure-identity>=1.25.2",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.1",
    "aiohttp>=3.11.0",
    "jinja2>=3.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/curate_common"]
